# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Recruitment Dashboard API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/requirements": {
      "get": {
        "tags": [
          "Requirements"
        ],
        "summary": "List Requirements",
        "operationId": "list_requirements_api_requirements_get",
        "parameters": [
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include inactive (closed) requirements in results.",
              "default": false,
              "title": "Include Inactive"
            },
            "description": "Include inactive (closed) requirements in results."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequirementRead"
                  },
                  "title": "Response List Requirements Api Requirements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Requirements"
        ],
        "summary": "Create Requirement",
        "operationId": "create_requirement_api_requirements_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/requirements/{requirement_id}": {
      "patch": {
        "tags": [
          "Requirements"
        ],
        "summary": "Update Requirement",
        "operationId": "update_requirement_api_requirements__requirement_id__patch",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Requirement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/requirements/{requirement_id}/status": {
      "patch": {
        "tags": [
          "Requirements"
        ],
        "summary": "Update Requirement Status",
        "operationId": "update_requirement_status_api_requirements__requirement_id__status_patch",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Requirement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/requirements/extract": {
      "post": {
        "tags": [
          "Requirements"
        ],
        "summary": "Extract Requirement",
        "operationId": "extract_requirement_api_requirements_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementExtractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matching/{requirement_id}": {
      "post": {
        "tags": [
          "Matching"
        ],
        "summary": "Run Matching",
        "operationId": "run_matching_api_matching__requirement_id__post",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Requirement Id"
            }
          },
          {
            "name": "match_all",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, run matching across all candidates (not only 'new' status)",
              "default": false,
              "title": "Match All"
            },
            "description": "If true, run matching across all candidates (not only 'new' status)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchResultRead"
                  },
                  "title": "Response Run Matching Api Matching  Requirement Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Matching"
        ],
        "summary": "Get Results",
        "operationId": "get_results_api_matching__requirement_id__get",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Requirement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchResultRead"
                  },
                  "title": "Response Get Results Api Matching  Requirement Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matching/{requirement_id}/candidates/{candidate_id}": {
      "post": {
        "tags": [
          "Matching"
        ],
        "summary": "Run Candidate Matching",
        "operationId": "run_candidate_matching_api_matching__requirement_id__candidates__candidate_id__post",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Requirement Id"
            }
          },
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchResultRead"
                  },
                  "title": "Response Run Candidate Matching Api Matching  Requirement Id  Candidates  Candidate Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matching/{requirement_id}/candidates/{candidate_id}/status": {
      "patch": {
        "tags": [
          "Matching"
        ],
        "summary": "Update Candidate Status",
        "operationId": "update_candidate_status_api_matching__requirement_id__candidates__candidate_id__status_patch",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Requirement Id"
            }
          },
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchStatusUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateRequirementStatusRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matching/{requirement_id}/bulk/reject-zero": {
      "post": {
        "tags": [
          "Matching"
        ],
        "summary": "Reject Zero Scores",
        "operationId": "reject_zero_scores_api_matching__requirement_id__bulk_reject_zero_post",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Requirement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkStatusUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matching/{requirement_id}/bulk/threshold": {
      "post": {
        "tags": [
          "Matching"
        ],
        "summary": "Apply Threshold Status",
        "operationId": "apply_threshold_status_api_matching__requirement_id__bulk_threshold_post",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Requirement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchThresholdStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkStatusUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matching/overview/{requirement_id}": {
      "get": {
        "tags": [
          "Matching"
        ],
        "summary": "Get Requirement Overview",
        "operationId": "get_requirement_overview_api_matching_overview__requirement_id__get",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Requirement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementOverviewRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkStatusUpdateResponse": {
        "properties": {
          "requirement_id": {
            "type": "integer",
            "title": "Requirement Id"
          },
          "updated_count": {
            "type": "integer",
            "title": "Updated Count"
          },
          "status": {
            "type": "string",
            "enum": [
              "processing",
              "rejected",
              "hired"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "requirement_id",
          "updated_count",
          "status"
        ],
        "title": "BulkStatusUpdateResponse"
      },
...
      "CandidateRequirementStatusRead": {
        "properties": {
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "requirement_id": {
            "type": "integer",
            "title": "Requirement Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "not_applied",
              "new",
              "processing",
              "rejected",
              "hired"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "requirement_id",
          "status"
        ],
        "title": "CandidateRequirementStatusRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchResultRead": {
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/CandidateMatchView"
          },
          "requirement": {
            "$ref": "#/components/schemas/RequirementMatchView"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "status": {
            "type": "string",
            "enum": [
              "not_applied",
              "new",
              "processing",
              "rejected",
              "hired"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "candidate",
          "requirement",
          "score",
          "reason",
          "status"
        ],
        "title": "MatchResultRead"
      },
      "MatchStatusUpdateRequest": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "not_applied",
              "new",
              "processing",
              "rejected",
              "hired"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "MatchStatusUpdateRequest"
      },
      "MatchThresholdStatusRequest": {
        "properties": {
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "status": {
            "type": "string",
            "enum": [
              "processing",
              "rejected",
              "hired"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "threshold",
          "status"
        ],
        "title": "MatchThresholdStatusRequest"
      },
...
      "RequirementCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/RequirementSkillInput"
            },
            "type": "array",
            "title": "Skills"
          },
          "required_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Skills"
          },
          "min_experience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Experience"
          },
          "max_experience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Experience"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "min_ctc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Ctc"
          },
          "max_ctc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Ctc"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "qualification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualification"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "RequirementCreate"
      },
      "RequirementExtractRequest": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 20,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "RequirementExtractRequest"
      },
      "RequirementExtractResponse": {
        "properties": {
          "requirement": {
            "$ref": "#/components/schemas/RequirementCreate"
          }
        },
        "type": "object",
        "required": [
          "requirement"
        ],
        "title": "RequirementExtractResponse"
      },
...
      "RequirementOverviewRead": {
        "properties": {
          "requirement_id": {
            "type": "integer",
            "title": "Requirement Id"
          },
          "total_current_candidates": {
            "type": "integer",
            "title": "Total Current Candidates"
          },
          "total_rejected_candidates": {
            "type": "integer",
            "title": "Total Rejected Candidates"
          },
          "total_hired_candidates": {
            "type": "integer",
            "title": "Total Hired Candidates"
          },
          "total_processing_candidates": {
            "type": "integer",
            "title": "Total Processing Candidates"
          }
        },
        "type": "object",
        "required": [
          "requirement_id",
          "total_current_candidates",
          "total_rejected_candidates",
          "total_hired_candidates",
          "total_processing_candidates"
        ],
        "title": "RequirementOverviewRead"
      },
      "RequirementRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/RequirementSkillRead"
            },
            "type": "array",
            "title": "Skills"
          },
          "required_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Skills"
          },
          "min_experience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Experience"
          },
          "max_experience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Experience"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "min_ctc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Ctc"
          },
          "max_ctc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Ctc"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "qualification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualification"
          },
          "summary_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Text"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "RequirementRead"
      },
...
      "RequirementSkillRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "min_experience_months": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Experience Months"
          },
          "min_experience_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Experience Years"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RequirementSkillRead"
      },
      "RequirementStatusUpdate": {
        "properties": {
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "is_active"
        ],
        "title": "RequirementStatusUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}